requirements.txt
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
idna==2.7
jedi==0.12.0
kafka-python==1.4.3
parso==0.2.1
pycodestyle==2.4.0
requests==2.19.1
urllib3==1.23
